Properly subclass PopupMenu.PopupSubMenuMenuItem using GObject.registerClass().
Gnome 3.34(.1) breaks the fix for Gnome 3.32 - seems to be an underlying
change on how PopupMenu.PopupSubMenuMenuItem is internally implemented.

Fix this by using GObject.registerClass(..) as described here:
   https://stackoverflow.com/a/57549355

This fixes the extension for Gnome 3.34(.1) and hopefully "futureproofs" this.

(Whatever these guys over at Gnome Foundation are smoking, it is
something good apparently.)